+++ src/components/com_named.ts
@@ -1,5 +1,5 @@
 import {Entity, Game} from "../game.js";
-import {Get, Has} from "./com_index.js";
+import {Has} from "./com_index.js";
 
 export interface Named {
     Name: string;
@@ -7,14 +7,14 @@ export interface Named {
 
 export function named(Name: string) {
     return (game: Game, entity: Entity) => {
-        game.World[entity] |= Has.Named;
-        game[Get.Named][entity] = <Named>{Name};
+        game.World.Mask[entity] |= Has.Named;
+        game.World.Named[entity] = <Named>{Name};
     };
 }
 
 export function find_first(game: Game, name: string) {
-    for (let i = 0; i < game[Get.Named].length; i++) {
-        let named = game[Get.Named][i];
+    for (let i = 0; i < game.World.Named.length; i++) {
+        let named = game.World.Named[i];
         if (named && named.Name === name) {
             return i;
         }