+++ src/components/com_shake.ts
@@ -1,5 +1,5 @@
 import {Entity, Game} from "../game.js";
-import {Get, Has} from "./com_index.js";
+import {Has} from "./com_index.js";
 
 export interface Shake {
     Duration: number;
@@ -10,8 +10,8 @@ export interface Shake {
  */
 export function shake(Duration = 0) {
     return (game: Game, entity: Entity) => {
-        game.World[entity] |= Has.Shake;
-        game[Get.Shake][entity] = <Shake>{
+        game.World.Mask[entity] |= Has.Shake;
+        game.World.Shake[entity] = <Shake>{
             Duration,
         };
     };