+++ src/systems/sys_collide.ts
@@ -1,5 +1,5 @@
 import {Collide} from "../components/com_collide.js";
-import {Get, Has} from "../components/com_index.js";
+import {Has} from "../components/com_index.js";
 import {Transform} from "../components/com_transform.js";
 import {Game} from "../game.js";
 import {Vec3} from "../math/index.js";
@@ -12,10 +12,10 @@ export function sys_collide(game: Game, delta: number) {
     // Collect all colliders.
     let static_colliders: Collide[] = [];
     let dynamic_colliders: Collide[] = [];
-    for (let i = 0; i < game.World.length; i++) {
-        if ((game.World[i] & QUERY) === QUERY) {
-            let transform = game[Get.Transform][i];
-            let collider = game[Get.Collide][i];
+    for (let i = 0; i < game.World.Mask.length; i++) {
+        if ((game.World.Mask[i] & QUERY) === QUERY) {
+            let transform = game.World.Transform[i];
+            let collider = game.World.Collide[i];
 
             // Prepare the collider for this tick's detection.
             collider.Collisions = [];