+++ src/systems/sys_control_player.ts
@@ -1,4 +1,4 @@
-import {Get, Has} from "../components/com_index.js";
+import {Has} from "../components/com_index.js";
 import {Entity, Game} from "../game.js";
 import {Vec3} from "../math/index.js";
 import {from_axis} from "../math/quat.js";
@@ -8,18 +8,18 @@ const AXIS_X = <Vec3>[1, 0, 0];
 const AXIS_Y = <Vec3>[0, 1, 0];
 
 export function sys_control_player(game: Game, delta: number) {
-    for (let i = 0; i < game.World.length; i++) {
-        if ((game.World[i] & QUERY) === QUERY) {
+    for (let i = 0; i < game.World.Mask.length; i++) {
+        if ((game.World.Mask[i] & QUERY) === QUERY) {
             update(game, i, delta);
         }
     }
 }
 
 function update(game: Game, entity: Entity, delta: number) {
-    let control = game[Get.PlayerControl][entity];
+    let control = game.World.PlayerControl[entity];
 
     if (control.Move) {
-        let move = game[Get.Move][entity];
+        let move = game.World.Move[entity];
         if (game.InputState.KeyW) {
             // Move forward
             move.Directions.push([0, 0, 1]);
@@ -39,7 +39,7 @@ function update(game: Game, entity: Entity, delta: number) {
     }
 
     if (control.Yaw) {
-        let move = game[Get.Move][entity];
+        let move = game.World.Move[entity];
         let yaw_delta = game.InputEvent.mouse_x * move.RotateSpeed * delta;
         if (yaw_delta !== 0) {
             move.Yaws.push(from_axis([0, 0, 0, 0], AXIS_Y, -yaw_delta));
@@ -47,7 +47,7 @@ function update(game: Game, entity: Entity, delta: number) {
     }
 
     if (control.Pitch) {
-        let move = game[Get.Move][entity];
+        let move = game.World.Move[entity];
         let pitch_delta = game.InputEvent.mouse_y * move.RotateSpeed * delta;
         if (pitch_delta !== 0) {
             move.Pitches.push(from_axis([0, 0, 0, 0], AXIS_X, pitch_delta));