+++ src/systems/sys_draw.ts
@@ -1,5 +1,5 @@
 import {DrawKind, DrawMarker} from "../components/com_draw.js";
-import {Get, Has} from "../components/com_index.js";
+import {Has} from "../components/com_index.js";
 import {Game} from "../game.js";
 import {Vec3} from "../math/index.js";
 import {get_translation} from "../math/mat4.js";
@@ -12,10 +12,10 @@ export function sys_draw(game: Game, delta: number) {
     game.Context2D.clearRect(0, 0, game.ViewportWidth, game.ViewportHeight);
     let position = <Vec3>[0, 0, 0];
 
-    for (let i = 0; i < game.World.length; i++) {
-        if ((game.World[i] & QUERY) == QUERY) {
+    for (let i = 0; i < game.World.Mask.length; i++) {
+        if ((game.World.Mask[i] & QUERY) == QUERY) {
             // World position.
-            get_translation(position, game[Get.Transform][i].World);
+            get_translation(position, game.World.Transform[i].World);
             // NDC position.
             transform_point(position, position, game.Cameras[0].PV);
 
@@ -28,7 +28,7 @@ export function sys_draw(game: Game, delta: number) {
                 0.5 * (-position[1] + 1) * game.ViewportHeight
             );
 
-            let draw = game[Get.Draw][i];
+            let draw = game.World.Draw[i];
             switch (draw.Kind) {
                 case DrawKind.Marker:
                     draw_marker(game, draw);