+++ src/systems/sys_draw2d.ts
@@ -1,5 +1,5 @@
 import {DrawKind, DrawRect} from "../components/com_draw.js";
-import {Get, Has} from "../components/com_index.js";
+import {Has} from "../components/com_index.js";
 import {Game} from "../game.js";
 
 const QUERY = Has.Transform2D | Has.Draw;
@@ -8,9 +8,9 @@ export function sys_draw2d(game: Game, delta: number) {
     game.Context2D.resetTransform();
     game.Context2D.clearRect(0, 0, game.ViewportWidth, game.ViewportHeight);
 
-    for (let i = 0; i < game.World.length; i++) {
-        if ((game.World[i] & QUERY) == QUERY) {
-            let transform = game[Get.Transform2D][i];
+    for (let i = 0; i < game.World.Mask.length; i++) {
+        if ((game.World.Mask[i] & QUERY) == QUERY) {
+            let transform = game.World.Transform2D[i];
             game.Context2D.setTransform(
                 transform.World[0],
                 transform.World[1],
@@ -20,7 +20,7 @@ export function sys_draw2d(game: Game, delta: number) {
                 transform.World[5]
             );
 
-            let draw = game[Get.Draw][i];
+            let draw = game.World.Draw[i];
             switch (draw.Kind) {
                 case DrawKind.Rect:
                     draw_rect(game, draw);