+++ src/systems/sys_light.ts
@@ -1,4 +1,4 @@
-import {Get, Has} from "../components/com_index.js";
+import {Has} from "../components/com_index.js";
 import {Entity, Game} from "../game.js";
 
 const QUERY = Has.Transform | Has.Light;
@@ -6,13 +6,13 @@ const QUERY = Has.Transform | Has.Light;
 export function sys_light(game: Game, delta: number) {
     game.Lights = [];
 
-    for (let i = 0; i < game.World.length; i++) {
-        if ((game.World[i] & QUERY) === QUERY) {
+    for (let i = 0; i < game.World.Mask.length; i++) {
+        if ((game.World.Mask[i] & QUERY) === QUERY) {
             update(game, i);
         }
     }
 }
 
 function update(game: Game, entity: Entity) {
-    game.Lights.push(game[Get.Light][entity]);
+    game.Lights.push(game.World.Light[entity]);
 }