+++ src/systems/sys_mimic.ts
@@ -1,4 +1,4 @@
-import {Get, Has} from "../components/com_index.js";
+import {Has} from "../components/com_index.js";
 import {Game} from "../game.js";
 import {get_rotation, get_translation} from "../math/mat4.js";
 import {slerp} from "../math/quat.js";
@@ -7,11 +7,11 @@ import {lerp} from "../math/vec3.js";
 const QUERY = Has.Transform | Has.Mimic;
 
 export function sys_mimic(game: Game, delta: number) {
-    for (let i = 0; i < game.World.length; i++) {
-        if ((game.World[i] & QUERY) === QUERY) {
-            let follower_transform = game[Get.Transform][i];
-            let follower_mimic = game[Get.Mimic][i];
-            let target_transform = game[Get.Transform][follower_mimic.target];
+    for (let i = 0; i < game.World.Mask.length; i++) {
+        if ((game.World.Mask[i] & QUERY) === QUERY) {
+            let follower_transform = game.World.Transform[i];
+            let follower_mimic = game.World.Mimic[i];
+            let target_transform = game.World.Transform[follower_mimic.target];
             let target_world_position = get_translation([0, 0, 0], target_transform.World);
             let target_world_rotation = get_rotation([0, 0, 0, 0], target_transform.World);
             // XXX Follower must be a top-level transform for this to work.